+++ src/captain/acmd/other.rs
@@ -13,3 +13,22 @@ use smash::app::*;
 use smash::phx::Vector3f;
 use crate::util::*;
 use super::*;
+pub fn install() {
+    smashline::install_acmd_scripts!(
+		captain_landing
+	);
+}
+
+
+
+#[acmd_script(
+    agent = "captain",
+    script =  "effect_landingheavy",
+    category = ACMD_EFFECT)]
+unsafe fn captain_landing(fighter: &mut L2CAgentBase) {
+    let lua_state = fighter.lua_state_agent;
+		frame(fighter.lua_state_agent, 2.0);
+		if macros::is_excute(fighter) {
+			macros::LANDING_EFFECT(fighter, Hash40::new("sys_landing_smoke"), Hash40::new("top"), 0, 0, 0, 0, 0, 0, 0.9, 0, 0, 0, 0, 0, 0, true);
+		}
+}
